requirements.txt
Django==4.2.7
Pillow==10.0.1
django-crispy-forms==2.0
crispy-tailwind==0.5.0
python-decouple==3.8
google-generativeai==0.3.1

# Production dependencies
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0